Fix SME palo alto vs. peninsula
Alan W's change:

Turns out that after tweaks to dancefinder so that SME would show up as the
Tuesday peninsula dance (last October), the generated BAERS dance URL was the
/peninsula page, which would send people to ASE.

I bit the bullet and created a series page for the BACDS/BAERS co-sponsored
dance, at /series/english/palo_alto_reg, copying over the content from
/peninsula/ and doing a series of tweaks until it looked right; may have missed
something.  Found that the March calendar entry had "REGENCY" rather than
"ENGLISH/REGENCY", so fixed that, and updated the pro-forma entries for August
and December, when that dance is on break, to reflect that.

Then edited dancefinder.pl to do a more-sophisticated mapping for 'SME', which
was formerly in the url-map of single events.  Now if it's Friday it's
palo_alto_reg, any other day it's peninsula, unless it's a special event with
its own url.