Add SUPPRESS option for "more info"
Checking in Alan W.'s change:

To get Ray's preferred text "For more info or to register"
rather than the script-generated "More Info", I put that text and the anchor
tag into the musicians field, put SUPPRESS in the url column, and modified
specialevents.pl to suppress the "More Info" if the dburl is "SUPPRESS".

And he suggested this as a TODO:

I'd like to suggest that in the new schema we add a way to, per entry,  put the
formatted output desired into one column and then just dump it out in
dancefinder or specialevents.  That would give us maximum flexibility for this
kind of request.  We could either add a column that does that, if ticked, or
just go by there being any value in that column.)