Tweak dancefinder.pl so berkeley can be Wed or Thu
Checking in Alan's change:

I changed dancefinder.pl so that it will recognize Wednesday or Thursday as
generating a /berkeley_wed URL.  (Didn't have to do this for the contra because
it didn't have to have code to disambiguate between two series at the same hall
as English does.  Terribly irregular and means the code has to know too much
about the data; this is the source of my request for the calendar database to
have a series-id for each entry, so we can simplify this code later.).